Merged PR 1192: Adopted @anpaz's suggestion for a serialization workflow.
This PR abstracts execution result and display data formatting into a new interface `IDisplaySerializer` that formats objects for displaying to clients. To demonstrate functionality that will be needed with IQ#, this PR also modifies the `moon-kernel` example to serialize all appropriate Lua objects to JSON. Using this mechanism, clients have full access to the structured output of kernels that opt-in to this behavior, making for a pretty generic way of interoperating different languages together.