Merged PR 1193: Enable display of arbitrary data, prototype tracer-style table support.
This PR builds on the serialization logic of !1192 to enable displaying arbitrary data in a uniform way between execution results and display channels. To demonstrate the new functionality, this PR also includes a new magic command `%version` in BaseEngine.cs that uses `TableSerializer` to present a table in a rich HTML format when available with graceful fallback to plain text tables. My hope is that this table serializer will be useful for displaying the outputs from tracer runs when applied to IQ#.

For instance:
```json
      "text/html": [
       "<table><thead><tr><th>Component</th><th>Version</th></tr></thead><tbody><tr><td>iecho</td><td>1.0.0.0</td></tr><tr><td>Jupyter Core</td><td>1.0.0.0</td></tr></tbody></table>"
      ],
      "text/plain": [
       "Component    Version\n",
       "------------ -------\n",
       "iecho        1.0.0.0\n",
       "Jupyter Core 1.0.0.0\n"
      ]
```